pdf_multimodel_call/requirements.txt
# Core LLM SDKs  (install only what you need)
google-genai>=1.5.0
anthropic>=0.30.0
openai>=1.40.0               # also covers Azure OpenAI

# File handling
python-dotenv>=1.0.0
PyMuPDF>=1.23.0              # PDF → text / image extraction (fitz)
Pillow>=10.0.0               # image handling

# Data processing & output
pandas>=2.1.0
openpyxl>=3.1.0              # Excel output (optional)
tenacity>=8.2.0              # retry logic

# Observability
structlog>=24.1.0            # structured JSON logging
prometheus-client>=0.20.0    # optional metrics endpoint
